# feat(bin): add single conn 1GB download benchmark