--- a/.github/workflows/bench.yml
+++ b/.github/workflows/bench.yml
@@ -73,7 +73,7 @@ jobs:
       # Pin the benchmark run to core 0 and run all benchmarks at elevated priority.
       - name: Run cargo bench
         run: |
-          nice -n -20 \
+          taskset -c 0 nice -n -20 \
             cargo "+$TOOLCHAIN" bench --features bench -- --noplot | tee results.txt
 
       # Compare various configurations of neqo against msquic, and gather perf data