--- a/.github/workflows/bench.yml
+++ b/.github/workflows/bench.yml
@@ -70,11 +70,19 @@ jobs:
       - name: Prepare machine
         run: sudo /root/bin/prep.sh
 
-      # Pin the benchmark run to core 0 and run all benchmarks at elevated priority.
       - name: Run cargo bench
         run: |
-          taskset -c 0 nice -n -20 \
-            cargo "+$TOOLCHAIN" bench --features bench -- --noplot | tee results.txt
+          # Pin all but neqo-bin benchmarks to CPU 0.
+          benchmarks=(
+            ("taskset -c 0" "--exclude neqo-bin")
+            ("" "--package neqo-bin")
+          )
+
+          for config in "${benchmarks[@]}"; do
+            read -r cpu_pinning packages <<< "${config}"
+            # run all benchmarks at elevated priority.
+            $cpu_pinning nice -n -20 cargo "+$TOOLCHAIN" bench $packages --features bench -- --noplot | tee -a results.txt
+          done
 
       # Compare various configurations of neqo against msquic, and gather perf data
       # during the hyperfine runs.