--- a/.github/workflows/bench.yml
+++ b/.github/workflows/bench.yml
@@ -74,14 +74,13 @@ jobs:
         run: |
           # Pin all but neqo-bin benchmarks to CPU 0. neqo-bin benchmarks run both a client and a server, thus benefiting from multiple CPU core.
           benchmarks=(
-            ("taskset -c 0" "--exclude neqo-bin")
-            ("" "--package neqo-bin")
+            "taskset -c 0,--exclude neqo-bin"
+            ",--package neqo-bin"
           )
-
           for config in "${benchmarks[@]}"; do
-            read -r cpu_pinning packages <<< "${config}"
+            IFS=',' read -r cpu_pinning packages <<< "$config"
             # Run all benchmarks at elevated priority.
-            $cpu_pinning nice -n -20 cargo "+$TOOLCHAIN" bench "$packages" --features bench -- --noplot | tee -a results.txt
+            nice -n -20 "$cpu_pinning" cargo "+$TOOLCHAIN" bench "$packages" --features bench -- --noplot | tee -a results.txt
           done
 
       # Compare various configurations of neqo against msquic, and gather perf data