--- a/neqo-bin/src/server/mod.rs
+++ b/neqo-bin/src/server/mod.rs
@@ -24,28 +24,28 @@ use futures::{
     future::{select, select_all, Either},
     FutureExt,
 };
-use neqo_bin::udp;
 use neqo_common::{hex, qinfo, qwarn, Datagram, Header};
 use neqo_crypto::{
     constants::{TLS_AES_128_GCM_SHA256, TLS_AES_256_GCM_SHA384, TLS_CHACHA20_POLY1305_SHA256},
     generate_ech_keys, init_db, random, AntiReplay, Cipher,
 };
 use neqo_http3::{
-    Error, Http3OrWebTransportStream, Http3Parameters, Http3Server, Http3ServerEvent, StreamId,
+    Http3OrWebTransportStream, Http3Parameters, Http3Server, Http3ServerEvent, StreamId,
 };
 use neqo_transport::{
     server::ValidateAddress, ConnectionIdGenerator, Output, RandomConnectionIdGenerator, Version,
 };
+use old_https::Http09Server;
 use tokio::time::Sleep;
 
-use crate::old_https::Http09Server;
+use crate::{udp, SharedArgs};
 
 const ANTI_REPLAY_WINDOW: Duration = Duration::from_secs(10);
 
 mod old_https;
 
 #[derive(Debug)]
-pub enum ServerError {
+pub enum Error {
     ArgumentError(&'static str),
     Http3Error(neqo_http3::Error),
     IoError(io::Error),
@@ -53,44 +53,44 @@ pub enum ServerError {
     TransportError(neqo_transport::Error),
 }
 
-impl From<io::Error> for ServerError {
+impl From<io::Error> for Error {
     fn from(err: io::Error) -> Self {
         Self::IoError(err)
     }
 }
 
-impl From<neqo_http3::Error> for ServerError {
+impl From<neqo_http3::Error> for Error {
     fn from(err: neqo_http3::Error) -> Self {
         Self::Http3Error(err)
     }
 }
 
-impl From<qlog::Error> for ServerError {
+impl From<qlog::Error> for Error {
     fn from(_err: qlog::Error) -> Self {
         Self::QlogError
     }
 }
 
-impl From<neqo_transport::Error> for ServerError {
+impl From<neqo_transport::Error> for Error {
     fn from(err: neqo_transport::Error) -> Self {
         Self::TransportError(err)
     }
 }
 
-impl Display for ServerError {
+impl Display for Error {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         write!(f, "Error: {self:?}")?;
         Ok(())
     }
 }
 
-impl std::error::Error for ServerError {}
+impl std::error::Error for Error {}
 
 #[derive(Debug, Parser)]
 #[command(author, version, about, long_about = None)]
-struct Args {
+pub struct Args {
     #[command(flatten)]
-    shared: neqo_bin::SharedArgs,
+    shared: SharedArgs,
 
     /// List of IP:port to listen on
     #[arg(default_value = "[::]:4433")]
@@ -113,6 +113,22 @@ struct Args {
     /// This generates a new set of ECH keys when it is invoked.
     /// The resulting configuration is printed to stdout in hexadecimal format.
     ech: bool,
+}
+
+#[cfg(feature = "bench")]
+impl Default for Args {
+    fn default() -> Self {
+        use std::str::FromStr;
+
+        Self {
+            shared: crate::SharedArgs::default(),
+            hosts: vec!["[::]:12345".to_string()],
+            db: PathBuf::from_str("../test-fixture/db").unwrap(),
+            key: "key".to_string(),
+            retry: false,
+            ech: false,
+        }
+    }
 }
 
 impl Args {
@@ -342,7 +358,7 @@ impl HttpServer for SimpleServer {
                             }
                         } else {
                             stream
-                                .cancel_fetch(Error::HttpRequestIncomplete.code())
+                                .cancel_fetch(neqo_http3::Error::HttpRequestIncomplete.code())
                                 .unwrap();
                             continue;
                         };
@@ -568,11 +584,9 @@ enum Ready {
     Timeout,
 }
 
-#[tokio::main]
-async fn main() -> Result<(), io::Error> {
+pub async fn server(mut args: Args) -> Result<(), io::Error> {
     const HQ_INTEROP: &str = "hq-interop";
 
-    let mut args = Args::parse();
     assert!(!args.key.is_empty(), "Need at least one key");
 
     init_db(args.db.clone());