--- a/neqo-bin/src/udp.rs
+++ b/neqo-bin/src/udp.rs
@@ -23,6 +23,8 @@ use tokio::io::Interest;
 const RECV_BUF_SIZE: usize = u16::MAX as usize;
 
 pub struct Socket {
+    #[allow(unknown_lints)] // available with Rust v1.75
+    #[allow(clippy::struct_field_names)]
     socket: tokio::net::UdpSocket,
     state: UdpSocketState,
     recv_buf: Vec<u8>,