Benchmark neqo-server -> neqo-client 1GB transfer using criterion. Simply wraps the `neqo-client` and `neqo-server` `main` functions.

```
main/single-connection  time:   [9.6643 s 11.252 s 12.778 s]
                        thrpt:  [80.136 MiB/s 91.008 MiB/s 105.96 MiB/s]
```

(Note this is bytes, not bits.)

---

It would be great to have **end-to-end** benchmarks (in CI)  to land performance optimizations like https://github.com/mozilla/neqo/pull/1741 with confidence.

This pull request is a **proof of concept using criterion**.

Benefits:

- Little boilerplate, written in Rust.
- Easy to extend with various benchmarks (e.g. multiple connections, handshakes per second, ...)
- Leverage features of criterion (though arguably the wrong tool for the job)
- Hook easily into existing [`bench.yml`](https://github.com/mozilla/neqo/blob/main/.github/workflows/bench.yml) workflow including its baseline comparison

Obviously there  are a million other ways to do this. 
- Using [hyperfine](https://github.com/sharkdp/hyperfine/) like it is done in https://github.com/mozilla/neqo/pull/1750/, thus agnostic to the binary being tested
- Building on top of [`upload_test.sh`](https://github.com/mozilla/neqo/blob/main/test/upload_test.sh) simulating an actual network (latency, packet drop, ...), i.e. not just loopback.
- Build our own platform, like
    - https://microsoft.github.io/msquic/
    - https://observablehq.com/@libp2p-workspace/performance-dashboard
- ...

What are people's thoughts? Are there long-term plans for neqo? Are there prominent examples worth taking inspiration from? Does Firefox / Mozilla already have some continuous benchmarking setup worth hooking into?