@mxinden 2024-03-20
Making some progress. Actual numbers are still preliminary.

```
Benchmarking single-request-1gb/client: Warming up for 3.0000 s
Warning: Unable to complete 10 samples in 5.0s. You may wish to increase target time to 76.6s.
single-request-1gb/client
                        time:   [8.3150 s 9.7354 s 11.247 s]
                        thrpt:  [91.047 MiB/s 105.18 MiB/s 123.15 MiB/s]
                 change:
                        time:   [-29.289% -14.840% +0.7625%] (p = 0.10 > 0.05)
                        thrpt:  [-0.7568% +17.426% +41.421%]
                        No change in performance detected.

Benchmarking single-request-1mb/client: Warming up for 3.0000 s
Warning: Unable to complete 10 samples in 5.0s. You may wish to increase target time to 8.8s or enable flat sampling.
single-request-1mb/client
                        time:   [159.50 ms 162.72 ms 167.38 ms]
                        thrpt:  [5.9743 MiB/s 6.1454 MiB/s 6.2697 MiB/s]
                 change:
                        time:   [+30.929% +53.168% +81.193%] (p = 0.00 < 0.05)
                        thrpt:  [-44.810% -34.712% -23.623%]
                        Performance has regressed.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) low severe

requests-per-second(10_000)/client
                        time:   [217.67 ms 242.75 ms 269.26 ms]
                        thrpt:  [37.140 Kelem/s 41.194 Kelem/s 45.942 Kelem/s]
                 change:
                        time:   [-13.591% -0.6218% +13.808%] (p = 0.92 > 0.05)
                        thrpt:  [-12.133% +0.6256% +15.729%]
                        No change in performance detected.

Benchmarking handshakes-per-second(100)/client: Warming up for 3.0000 s
Warning: Unable to complete 10 samples in 5.0s. You may wish to increase target time to 25.0s.
handshakes-per-second(100)/client
                        time:   [1.8661 s 1.9559 s 2.0763 s]
                        thrpt:  [48.162  elem/s 51.127  elem/s 53.587  elem/s]
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high mild
```