@larseggert 2024-03-25
Is there a way to group the benchmarks, so we can pin those where it makes sense to a core, and let others such as this one be  scheduled on multiple cores?